+++ vite.config.js
@@ -14,8 +14,8 @@ export default defineConfig({
             ],
             refresh: true,
             fonts: [
-                bunny('Instrument Sans', {
-                    weights: [400, 500, 600],
+                bunny('Outfit', {
+                    weights: [400, 500, 600, 700],
                 }),
             ],
         }),